+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.scom/syspack/pakages/tree/main) (0.0.x)
+ - bug with parsing hidden (and non package directories) in spack (0.0.19)
  - bugfixes to install and adding GitHub action (0.0.18)
    - checking for existing mirrors before adding blindly
  - updates to spack build (0.0.17)